update the interface to double values

+++ sdk/master_board_sdk/example/example.cpp
@@ -110,7 +110,7 @@ int main(int argc, char **argv)
 							cur = iq_sat;
 						if (cur < -iq_sat)
 							cur = -iq_sat;
-						robot_if.motors[i].SetCurrentReference(static_cast<float>(cur));
+						robot_if.motors[i].SetCurrentReference(cur);
 					}
 				}
 				break;
